Controls&DisplayUserAdded

+++ Hajs_Sie_Zgadza/src/hajs_sie_zgadza/User.java
@@ -4,9 +4,9 @@ public class User {
     
     private int idUser;
     private String userName;
-    private int cash;
+    private double cash;
     
-    public User(int idUser, String userName,int cash) {
+    public User(int idUser, String userName,double cash) {
         
         this.idUser = idUser;
         this.userName = userName;
@@ -18,7 +18,7 @@ public class User {
         return userName;
     }
     
-    public int getCash() {
+    public double getCash() {
         return cash;
     }
     
@@ -26,7 +26,7 @@ public class User {
         this.userName = userName;
     }
     
-    public void setCash(int cash) {
+    public void setCash(double cash) {
         this.cash = cash;
     }
